chore(general): Move RXValidator to StatusQ
