+++ storybook/CMakeLists.txt
@@ -36,6 +36,7 @@ set(STATUSQ_HEADERS
     ${STATUSQ_DIR}/include/StatusQ/typesregistration.h
     ${STATUSQ_DIR}/include/StatusQ/QClipboardProxy.h
     ${STATUSQ_DIR}/include/StatusQ/statussyntaxhighlighter.h
+    ${STATUSQ_DIR}/include/StatusQ/rxvalidator.h
 )
 
 set(STATUSQ_SOURCES
@@ -43,6 +44,7 @@ set(STATUSQ_SOURCES
     ${STATUSQ_DIR}/src/typesregistration.cpp
     ${STATUSQ_DIR}/src/QClipboardProxy.cpp
     ${STATUSQ_DIR}/src/statussyntaxhighlighter.cpp
+    ${STATUSQ_DIR}/src/rxvalidator.cpp
 )
 
 if(APPLE)
