+++ ui/StatusQ/sandbox/CMakeLists.txt
@@ -17,6 +17,7 @@ set(STATUSQ_HEADERS
     ${STATUSQ_DIR}/include/StatusQ/typesregistration.h
     ${STATUSQ_DIR}/include/StatusQ/QClipboardProxy.h
     ${STATUSQ_DIR}/include/StatusQ/statussyntaxhighlighter.h
+    ${STATUSQ_DIR}/include/StatusQ/rxvalidator.h
 )
 
 set(STATUSQ_SOURCES
@@ -24,6 +25,7 @@ set(STATUSQ_SOURCES
     ${STATUSQ_DIR}/src/typesregistration.cpp
     ${STATUSQ_DIR}/src/QClipboardProxy.cpp
     ${STATUSQ_DIR}/src/statussyntaxhighlighter.cpp
+    ${STATUSQ_DIR}/src/rxvalidator.cpp
 )
 
 if(APPLE)
